+++ CHANGELOG.md
@@ -46,6 +46,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added optional `embedded_io::Write` impl for `Vec`.
 
+### Added
+
+- Added `from_vec` and `into_sorted_vec` implementations for `BinaryHeap`.
+
 ### Changed
 
 - `bytes::BufMut` is now implemented on `VecInner`.